requirements.txt
openai>=1.3.9
jinja2
python-dotenv
yfinance
requests
pandas
fredapi
streamlit
plotly
pyyaml
lxml
plotly
alpha_vantage
requests
setuptools>=67
pandas_datareader>=0.10
numpy
scipy